fix: duplicated dark mode switch and share page button
