+++ package.json
@@ -90,7 +90,6 @@
     "gatsby-transformer-yaml": "^2.4.1",
     "gh-pages": "^2.2.0",
     "graphql": "^15.0.0",
-    "instant.page": "^3.0.0",
     "lodash": "^4.17.15",
     "moment-timezone": "^0.5.28",
     "node-fetch": "^2.6.0",
