+++ src/components/Navigation/Navigation.Header.tsx
@@ -8,12 +8,10 @@ import Section from '@components/Section';
 import Logo from '@components/Logo';
 import useIsDarkMode from '@hooks/useIsDark';
 import mediaqueries from '@styles/media';
-import { IconChevronLeft, IconEx, IconLink } from '@components/Icons';
-import {
-  copyToClipboard,
-  getWindowDimensions,
-  getBreakpointFromTheme,
-} from '@utils';
+import { IconChevronLeft, IconEx } from '@components/Icons';
+import { getWindowDimensions, getBreakpointFromTheme } from '@utils';
+import ShareButton from '@components/ShareButton';
+import ModeSwitch from '../ModeSwitch';
 
 const strNavToHome = '回到主页';
 
@@ -92,8 +90,8 @@ function NavigationHeader({ location }: NavigationHeaderPropos) {
               </button>
             ) : (
               <>
-                <SharePageButton />
-                <DarkModeToggle />
+                <ShareButton />
+                <ModeSwitch />
               </>
             )}
           </NavControls>
@@ -104,74 +102,6 @@ function NavigationHeader({ location }: NavigationHeaderPropos) {
 }
 
 export default NavigationHeader;
-
-const strActiveLightMode = '打开 Light Mode';
-const strActiveDarkMode = '打开 Dark Mode';
-
-function DarkModeToggle() {
-  const [colorMode, setColorMode] = useColorMode();
-  const isDark = colorMode === `dark`;
-
-  const toggleColorMode = useCallback(
-    (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-      event.preventDefault();
-      setColorMode(isDark ? `default` : `dark`);
-    },
-    [isDark, setColorMode]
-  );
-
-  return (
-    <IconWrapper
-      isDark={isDark}
-      onClick={toggleColorMode}
-      data-a11y="false"
-      aria-label={isDark ? strActiveLightMode : strActiveDarkMode}
-      title={isDark ? strActiveLightMode : strActiveDarkMode}
-    >
-      <MoonOrSun isDark={isDark} />
-      <MoonMask isDark={isDark} />
-    </IconWrapper>
-  );
-}
-
-const strCopyUrlToClipboard = '复制 URL 到剪切板';
-
-function SharePageButton() {
-  const [hasCopied, setHasCopied] = useState<boolean>(false);
-  const [colorMode] = useColorMode();
-  const isDark = colorMode === `dark`;
-  const fill = isDark ? '#fff' : '#000';
-
-  const copyToClipboardOnClick = useCallback(() => {
-    if (hasCopied) return;
-
-    copyToClipboard(window.location.href);
-    setHasCopied(true);
-
-    setTimeout(() => {
-      setHasCopied(false);
-    }, 1000);
-  }, [hasCopied]);
-
-  return (
-    <IconWrapper
-      isDark={isDark}
-      onClick={copyToClipboardOnClick}
-      data-a11y="false"
-      aria-label={strCopyUrlToClipboard}
-      title={strCopyUrlToClipboard}
-    >
-      <IconLink fill={fill} />
-      <ToolTip
-        isDark={isDark}
-        hasCopied={hasCopied}
-        id="SharePageButtonTooltip"
-      >
-        复制成功
-      </ToolTip>
-    </IconWrapper>
-  );
-}
 
 const RootContainer = styled.div<{ detached: boolean; isDark: boolean }>`
   position: sticky;
@@ -292,140 +222,6 @@ const NavControls = styled.div`
   ${mediaqueries.phablet`
     right: -5px;
   `}
-`;
-
-const ToolTip = styled.div<{ isDark: boolean; hasCopied: boolean }>`
-  position: absolute;
-  padding: 4px 13px;
-  background: ${p => (p.isDark ? '#000' : 'rgba(0,0,0,0.1)')};
-  color: ${p => (p.isDark ? '#fff' : '#000')};
-  border-radius: 5px;
-  font-size: 14px;
-  top: 35px;
-  opacity: ${p => (p.hasCopied ? 1 : 0)};
-  transform: ${p => (p.hasCopied ? 'translateY(-3px)' : 'none')};
-  transition: transform 0.3s ease-in-out, opacity 0.3s ease-in-out;
-  white-space: nowrap;
-
-  &::before {
-    content: '';
-    position: absolute;
-    left: 0;
-    right: 0;
-    top: -6px;
-    margin: 0 auto;
-    width: 0;
-    height: 0;
-    border-left: 6px solid transparent;
-    border-right: 6px solid transparent;
-    border-bottom: 6px solid ${p => (p.isDark ? '#000' : 'rgba(0,0,0,0.1)')};
-  }
-`;
-
-const IconWrapper = styled.button<{ isDark: boolean }>`
-  opacity: 0.5;
-  position: relative;
-  border-radius: 5px;
-  width: 40px;
-  height: 25px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: opacity 0.3s ease;
-  margin-left: 30px;
-
-  &:hover {
-    opacity: 1;
-  }
-
-  &[data-a11y='true']:focus::after {
-    content: '';
-    position: absolute;
-    left: 0;
-    top: -30%;
-    width: 100%;
-    height: 160%;
-    border: 2px solid ${p => p.theme.colors.accent};
-    background: rgba(255, 255, 255, 0.01);
-    border-radius: 5px;
-  }
-
-  ${mediaqueries.tablet`
-    display: inline-flex;
-    transform: scale(0.708);
-    margin-left: 10px;
-
-    &:hover {
-      opacity: 0.5;
-    }
-  `}
-`;
-
-// This is based off a codepen! Much appreciated to: https://codepen.io/aaroniker/pen/KGpXZo
-const MoonOrSun = styled.div<{ isDark: boolean }>`
-  position: relative;
-  width: 24px;
-  height: 24px;
-  border-radius: 50%;
-  border: ${p => (p.isDark ? '4px' : '2px')} solid
-    ${p => p.theme.colors.primary};
-  background: ${p => p.theme.colors.primary};
-  transform: scale(${p => (p.isDark ? 0.55 : 1)});
-  transition: all 0.45s ease;
-  overflow: ${p => (p.isDark ? 'visible' : 'hidden')};
-
-  &::before {
-    content: '';
-    position: absolute;
-    right: -9px;
-    top: -9px;
-    height: 24px;
-    width: 24px;
-    border: 2px solid ${p => p.theme.colors.primary};
-    border-radius: 50%;
-    transform: translate(${p => (p.isDark ? '14px, -14px' : '0, 0')});
-    opacity: ${p => (p.isDark ? 0 : 1)};
-    transition: transform 0.45s ease;
-  }
-
-  &::after {
-    content: '';
-    width: 8px;
-    height: 8px;
-    border-radius: 50%;
-    margin: -4px 0 0 -4px;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    box-shadow: 0 -23px 0 ${p => p.theme.colors.primary},
-      0 23px 0 ${p => p.theme.colors.primary},
-      23px 0 0 ${p => p.theme.colors.primary},
-      -23px 0 0 ${p => p.theme.colors.primary},
-      15px 15px 0 ${p => p.theme.colors.primary},
-      -15px 15px 0 ${p => p.theme.colors.primary},
-      15px -15px 0 ${p => p.theme.colors.primary},
-      -15px -15px 0 ${p => p.theme.colors.primary};
-    transform: scale(${p => (p.isDark ? 1 : 0)});
-    transition: all 0.35s ease;
-
-    ${p => mediaqueries.tablet`
-      transform: scale(${p.isDark ? 0.92 : 0});
-    `}
-  }
-`;
-
-const MoonMask = styled.div<{ isDark: boolean }>`
-  position: absolute;
-  right: -1px;
-  top: -8px;
-  height: 24px;
-  width: 24px;
-  border-radius: 50%;
-  border: 0;
-  background: ${p => p.theme.colors.background};
-  transform: translate(${p => (p.isDark ? '14px, -14px' : '0, 0')});
-  opacity: ${p => (p.isDark ? 0 : 1)};
-  transition: ${p => p.theme.colorModeTransition}, transform 0.45s ease;
 `;
 
 const Hidden = styled.span`
