+++ src/sections/article/Article.Controls.tsx
@@ -1,236 +1,22 @@
-import React, { useState, useCallback } from 'react';
+import React from 'react';
 import styled from '@emotion/styled';
-import { useColorMode } from 'theme-ui';
 
-import mediaqueries from '@styles/media';
-import { copyToClipboard } from '@utils';
+import ShareButton from '@components/ShareButton';
+import ModeSwitch from '@components/ModeSwitch';
 
 function ArticleControls() {
   return (
     <NavControls>
-      <SharePageButton />
-      <DarkModeToggle />
+      <ShareButton />
+      <ModeSwitch />
     </NavControls>
   );
 }
 
 export default React.memo(ArticleControls);
 
-function DarkModeToggle() {
-  const [colorMode, setColorMode] = useColorMode();
-  const isDark = colorMode === `dark`;
-
-  const toggleColorMode = useCallback(
-    (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-      event.preventDefault();
-      setColorMode(isDark ? `default` : `dark`);
-    },
-    [isDark, setColorMode]
-  );
-
-  return (
-    <IconWrapper
-      onClick={toggleColorMode}
-      aria-label="Toggle dark and light mode"
-    >
-      <MoonOrSun isDark={isDark} />
-      <MoonMask isDark={isDark} />
-    </IconWrapper>
-  );
-}
-
-const ShareDarkModeOffIcon = () => (
-  <svg
-    width="24"
-    height="20"
-    viewBox="0 0 24 20"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M2 5C2 3.34328 3.34328 2 5 2H14C15.6567 2 17 3.34328 17 5V9C17 10.6567 15.6567 12 14 12H10C9.44771 12 9 12.4477 9 13C9 13.5523 9.44771 14 10 14H14C16.7613 14 19 11.7613 19 9V5C19 2.23872 16.7613 0 14 0H5C2.23872 0 0 2.23872 0 5V9C0 10.4938 0.656313 11.8361 1.6935 12.7509C2.10768 13.1163 2.73961 13.0767 3.10494 12.6625C3.47028 12.2483 3.43068 11.6164 3.0165 11.2511C2.39169 10.6999 2 9.89621 2 9V5ZM7 11C7 9.34328 8.34328 8 10 8H14C14.5523 8 15 7.55228 15 7C15 6.44772 14.5523 6 14 6H10C7.23872 6 5 8.23872 5 11V15C5 17.7613 7.23872 20 10 20H19C21.7613 20 24 17.7613 24 15V11C24 9.50621 23.3437 8.16393 22.3065 7.24906C21.8923 6.88372 21.2604 6.92332 20.8951 7.3375C20.5297 7.75168 20.5693 8.38361 20.9835 8.74894C21.6083 9.30007 22 10.1038 22 11V15C22 16.6567 20.6567 18 19 18H10C8.34328 18 7 16.6567 7 15V11Z"
-      fill="white"
-    />
-  </svg>
-);
-
-const ShareDarkModeOnIcon = () => (
-  <svg
-    width="24"
-    height="20"
-    viewBox="0 0 24 20"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M2 5C2 3.34328 3.34328 2 5 2H14C15.6567 2 17 3.34328 17 5V9C17 10.6567 15.6567 12 14 12H10C9.44771 12 9 12.4477 9 13C9 13.5523 9.44771 14 10 14H14C16.7613 14 19 11.7613 19 9V5C19 2.23872 16.7613 0 14 0H5C2.23872 0 0 2.23872 0 5V9C0 10.4938 0.656313 11.8361 1.6935 12.7509C2.10768 13.1163 2.73961 13.0767 3.10494 12.6625C3.47028 12.2483 3.43068 11.6164 3.0165 11.2511C2.39169 10.6999 2 9.89621 2 9V5ZM7 11C7 9.34328 8.34328 8 10 8H14C14.5523 8 15 7.55228 15 7C15 6.44772 14.5523 6 14 6H10C7.23872 6 5 8.23872 5 11V15C5 17.7613 7.23872 20 10 20H19C21.7613 20 24 17.7613 24 15V11C24 9.50621 23.3437 8.16393 22.3065 7.24906C21.8923 6.88372 21.2604 6.92332 20.8951 7.3375C20.5297 7.75168 20.5693 8.38361 20.9835 8.74894C21.6083 9.30007 22 10.1038 22 11V15C22 16.6567 20.6567 18 19 18H10C8.34328 18 7 16.6567 7 15V11Z"
-      fill="black"
-    />
-  </svg>
-);
-
-function SharePageButton() {
-  const [hasCopied, setHasCopied] = useState<boolean>(false);
-  const [colorMode] = useColorMode();
-  const isDark = colorMode === `dark`;
-
-  const copyToClipboardOnClick = useCallback(() => {
-    if (hasCopied) return;
-
-    copyToClipboard(window.location.href);
-    setHasCopied(true);
-
-    setTimeout(() => {
-      setHasCopied(false);
-    }, 1000);
-  }, [hasCopied]);
-
-  const Icon = isDark ? ShareDarkModeOffIcon : ShareDarkModeOnIcon;
-
-  return (
-    <IconWrapper
-      onClick={copyToClipboardOnClick}
-      data-a11y="false"
-      aria-label="Copy URL to clipboard"
-    >
-      <Icon />
-      <ToolTip isDark={isDark} hasCopied={hasCopied}>
-        复制成功
-      </ToolTip>
-    </IconWrapper>
-  );
-}
-
 const NavControls = styled.div`
   display: flex;
   align-items: center;
   justify-content: center;
-`;
-
-const ToolTip = styled.div<{ isDark: boolean; hasCopied: boolean }>`
-  position: absolute;
-  padding: 4px 13px;
-  background: ${p => (p.isDark ? '#000' : 'rgba(0,0,0,0.1)')};
-  color: ${p => (p.isDark ? '#fff' : '#000')};
-  border-radius: 5px;
-  font-size: 14px;
-  top: -35px;
-  opacity: ${p => (p.hasCopied ? 1 : 0)};
-  transform: ${p => (p.hasCopied ? 'translateY(-3px)' : 'none')};
-  transition: transform 0.3s ease-in-out, opacity 0.3s ease-in-out;
-  white-space: nowrap;
-
-  &::after {
-    content: '';
-    position: absolute;
-    left: 0;
-    right: 0;
-    bottom: -6px;
-    margin: 0 auto;
-    width: 0;
-    height: 0;
-    border-left: 6px solid transparent;
-    border-right: 6px solid transparent;
-    border-top: 6px solid ${p => (p.isDark ? '#000' : 'rgba(0,0,0,0.1)')};
-  }
-`;
-
-const IconWrapper = styled.button`
-  opacity: 0.5;
-  position: relative;
-  z-index: 200000000;
-  border-radius: 5px;
-  width: 40px;
-  height: 25px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: opacity 0.3s ease;
-  margin-left: 30px;
-
-  &:hover {
-    opacity: 1;
-  }
-
-  ${mediaqueries.tablet`
-    display: inline-flex;
-    transform: scale(0.9);
-    margin: 0 15px;
-
-
-    &:hover {
-      opacity: 0.5;
-    }
-  `}
-`;
-
-// This is based off a codepen! Much appreciated to: https://codepen.io/aaroniker/pen/KGpXZo
-const MoonOrSun = styled.div<{ isDark: boolean }>`
-  position: relative;
-  width: 24px;
-  height: 24px;
-  border-radius: 50%;
-  border: ${p => (p.isDark ? '4px' : '2px')} solid
-    ${p => p.theme.colors.primary};
-  background: ${p => p.theme.colors.primary};
-  transform: scale(${p => (p.isDark ? 0.55 : 1)});
-  transition: all 0.45s ease;
-  overflow: ${p => (p.isDark ? 'visible' : 'hidden')};
-
-  &::before {
-    content: '';
-    position: absolute;
-    right: -9px;
-    top: -9px;
-    height: 24px;
-    width: 24px;
-    border: 2px solid ${p => p.theme.colors.primary};
-    border-radius: 50%;
-    transform: translate(${p => (p.isDark ? '14px, -14px' : '0, 0')});
-    opacity: ${p => (p.isDark ? 0 : 1)};
-    transition: transform 0.45s ease;
-  }
-
-  &::after {
-    content: '';
-    width: 8px;
-    height: 8px;
-    border-radius: 50%;
-    margin: -4px 0 0 -4px;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    box-shadow: 0 -23px 0 ${p => p.theme.colors.primary},
-      0 23px 0 ${p => p.theme.colors.primary},
-      23px 0 0 ${p => p.theme.colors.primary},
-      -23px 0 0 ${p => p.theme.colors.primary},
-      15px 15px 0 ${p => p.theme.colors.primary},
-      -15px 15px 0 ${p => p.theme.colors.primary},
-      15px -15px 0 ${p => p.theme.colors.primary},
-      -15px -15px 0 ${p => p.theme.colors.primary};
-    transform: scale(${p => (p.isDark ? 1 : 0)});
-    transition: all 0.35s ease;
-
-    ${p => mediaqueries.tablet`
-      transform: scale(${p.isDark ? 0.92 : 0});
-    `}
-  }
-`;
-
-const MoonMask = styled.div<{ isDark: boolean }>`
-  position: absolute;
-  right: -1px;
-  top: -8px;
-  height: 24px;
-  width: 24px;
-  border-radius: 50%;
-  border: 0;
-  background: ${p => p.theme.colors.background};
-  transform: translate(${p => (p.isDark ? '14px, -14px' : '0, 0')});
-  opacity: ${p => (p.isDark ? 0 : 1)};
-  transition: transform 0.45s ease, ${p => p.theme.colorModeTransition};
 `;
